rename FloatMenu to Popover #50
