+++ components/index.web.tsx
@@ -8,7 +8,7 @@ export { default as Checkbox } from './checkbox/index.web';
 export { default as DatePicker } from './date-picker/index.web';
 export { default as Drawer } from './drawer/index.web';
 export { default as Flex } from './flex/index.web';
-export { default as FloatMenu } from './float-menu/index.web';
+export { default as Popover } from './popover/index.web';
 
 // export { default as Form } from './form/index.web';
 export { default as Grid } from './grid/index.web';
